pyproject.toml
[tool.poetry]
name = "misor"
version = "0.1.0"
description = "Build your MIS in no time."
authors = ["Rashid <rashidstanikzai.net@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
